rcar-gen3: u-boot: Support deploying multiple u-boot binaries
This commit updates recipes to support deploying multiple u-boot
binaries of Salvator-XS, H3ULCB and Ebisu boards. It also removed
uboot-control.inc which is no longer used.

+++ meta-rcar-gen3/recipes-bsp/u-boot/u-boot_2018.09.bb
@@ -1,6 +1,5 @@
 require u-boot-common_${PV}.inc
 require u-boot.inc
-require include/uboot-control.inc
 
 DEPENDS += "bc-native dtc-native"
 
@@ -11,17 +10,33 @@ SRC_URI = "${UBOOT_URL};branch=${BRANCH}"
 SRCREV = "70206a1b84e6e35c33b3760fae6a2a6dbe6ce534"
 PV = "v2018.09+git${SRCPV}"
 
-# Add u-boot options for H3
-EXTRA_OEMAKE_append_r8a7795 = " ${@get_uboot_config_opt(d)}"
-
-UBOOT_SREC ?= "u-boot-elf.srec"
-UBOOT_SREC_IMAGE ?= "u-boot-elf-${MACHINE}-${PV}-${PR}.srec"
-UBOOT_SREC_SYMLINK ?= "u-boot-elf-${MACHINE}.srec"
+UBOOT_SREC_SUFFIX = "srec"
+UBOOT_SREC ?= "u-boot-elf.${UBOOT_SREC_SUFFIX}"
+UBOOT_SREC_IMAGE ?= "u-boot-elf-${MACHINE}-${PV}-${PR}.${UBOOT_SREC_SUFFIX}"
+UBOOT_SREC_SYMLINK ?= "u-boot-elf-${MACHINE}.${UBOOT_SREC_SUFFIX}"
 
 do_deploy_append() {
-    install -m 644 ${B}/${UBOOT_SREC} ${DEPLOYDIR}/${UBOOT_SREC_IMAGE}
-    cd ${DEPLOYDIR}
-    rm -f ${UBOOT_SREC} ${UBOOT_SREC_SYMLINK}
-    ln -sf ${UBOOT_SREC_IMAGE} ${UBOOT_SREC_SYMLINK}
-    ln -sf ${UBOOT_SREC_IMAGE} ${UBOOT_SREC}
+    if [ -n "${UBOOT_CONFIG}" ]
+    then
+        for config in ${UBOOT_MACHINE}; do
+            i=$(expr $i + 1);
+            for type in ${UBOOT_CONFIG}; do
+                j=$(expr $j + 1);
+                if [ $j -eq $i ]
+                then
+                    install -m 644 ${B}/${config}/${UBOOT_SREC} ${DEPLOYDIR}/u-boot-elf-${type}-${PV}-${PR}.${UBOOT_SREC_SUFFIX}
+                    cd ${DEPLOYDIR}
+                    ln -sf u-boot-elf-${type}-${PV}-${PR}.${UBOOT_SREC_SUFFIX} u-boot-elf-${type}.${UBOOT_SREC_SUFFIX}
+                fi
+            done
+            unset j
+        done
+        unset i
+    else
+        install -m 644 ${B}/${UBOOT_SREC} ${DEPLOYDIR}/${UBOOT_SREC_IMAGE}
+        cd ${DEPLOYDIR}
+        rm -f ${UBOOT_SREC} ${UBOOT_SREC_SYMLINK}
+        ln -sf ${UBOOT_SREC_IMAGE} ${UBOOT_SREC_SYMLINK}
+        ln -sf ${UBOOT_SREC_IMAGE} ${UBOOT_SREC}
+    fi
 }
